Completed button onClick functions
this.props contains an object with count, and both increment
and decrement functions.

When + button is clicked, this.props.increment function is invoked
When - button is clicked, this.props.decrement function is invoked

When 'IncrementIfOdd' is clicked, first will check if count is odd. If
it is, then increment function is invoked. If even, nothing will happen.

Increment async button calls increment function after a delay of 1000ms.